organize rewriters
