+++ proxychain/responsemodifers/rewriters/css_resource_url_rewriter.go
@@ -1,0 +1,3 @@
+package rewriters
+
+// todo: implement
